deploy: dev-> main

+++ src/meeting/meeting.gateway.ts
@@ -8,7 +8,7 @@ import {
 } from '@nestjs/websockets'
 import { UseGuards } from '@nestjs/common'
 import { Server, Socket } from 'socket.io'
-import { OpenViduService } from './services/meeting.service'
+import { MeetingService } from './services/meeting.service'
 import { QueueService } from './services/queue.service'
 import { ConfigService } from '@nestjs/config'
 import { JwtAuthWsGuard } from '../guards/jwt-auth.ws.guard'
@@ -34,33 +34,38 @@ export class MeetingGateway
   private roomid: Map<string, string> = new Map()
   private isDevelopment: boolean
   constructor(
-    private readonly openviduService: OpenViduService,
+    private readonly meetingService: MeetingService,
     private readonly queueService: QueueService,
     private readonly configService: ConfigService,
   ) {
     this.isDevelopment = this.configService.get<string>('NODE_ENV') === 'dev'
   }
-  private connectedUsers: { [nickname: string]: string } = {} // nickname: socketId 형태로 변경
+  private connectedUsers: { [nickname: string]: Socket } = {} // nickname: socketId 형태로 변경
   private connectedSockets: { [socketId: string]: string } = {} // socketId: nickname 형태로 변경
   private cupidFlag: Map<string, boolean> = new Map()
+  private timerFlag: Map<string, boolean> = new Map()
+  private lastCupidFlag: Map<string,boolean> = new Map()
+
+  private acceptanceStatus: Record<string, boolean> = {}
   afterInit(server: Server) {
-    this.openviduService.server = server
+    this.meetingService.server = server
     console.log('WebSocket initialized')
   }
 
   handleConnection(client: Socket) {}
 
   handleDisconnect(client: Socket) {
-    const sessions = this.openviduService.getSessions()
+    const sessions = this.meetingService.getSessions()
     const participantName = this.connectedSockets[client.id]
-    if (!sessions.length) {
+    const user = client['user']
+    if (!sessions.length && user) {
       const gender = client['user'].gender
       this.queueService.removeParticipant(participantName, gender)
     }
 
     for (const sessionName in sessions) {
       if (sessions.hasOwnProperty(sessionName)) {
-        this.openviduService.removeParticipant(
+        this.meetingService.removeParticipant(
           sessionName,
           client,
           participantName,
@@ -70,6 +75,11 @@ export class MeetingGateway
     delete this.connectedSockets[client.id]
     delete this.connectedUsers[participantName]
     this.roomid.delete(participantName)
+
+    const nickname = this.connectedSockets[client.id]
+    if (nickname in this.acceptanceStatus) {
+      delete this.acceptanceStatus[nickname]
+    }
   }
 
   @SubscribeMessage('ready')
@@ -87,12 +97,10 @@ export class MeetingGateway
         participantName = payload.participantName
         gender = payload.gender
       }
-      // const participantName = client['user'].nickname
-      // const gender = client['user'].gender
 
       const existingSessionName = this.roomid.get(participantName)
       if (existingSessionName) {
-        this.openviduService.removeParticipant(
+        this.meetingService.removeParticipant(
           existingSessionName,
           client,
           participantName,
@@ -110,7 +118,7 @@ export class MeetingGateway
           this.roomid.set(female.name, sessionName)
         })
       }
-      this.connectedUsers[participantName] = client.id
+      this.connectedUsers[participantName] = client
       this.connectedSockets[client.id] = participantName
     } catch (error) {
       console.log('Error handling join Queue request:', error)
@@ -122,7 +130,7 @@ export class MeetingGateway
     client: Socket,
     payload: { participantName: string; gender: string },
   ) {
-    const sessions = this.openviduService.getSessions()
+    const sessions = this.meetingService.getSessions()
     let participantName
     let gender
     if (this.isDevelopment) {
@@ -137,7 +145,7 @@ export class MeetingGateway
 
     for (const sessionName in sessions) {
       if (sessions.hasOwnProperty(sessionName)) {
-        this.openviduService.removeParticipant(
+        this.meetingService.removeParticipant(
           sessionName,
           client,
           participantName,
@@ -151,17 +159,19 @@ export class MeetingGateway
 
   @SubscribeMessage('choose')
   handleChoose(client: Socket, payload: { sender: string; receiver: string }) {
+    // 해당 소켓이 존재하는 방을 찾기 위함
     const sessionName = this.roomid.get(payload.sender)
     if (sessionName) {
-      this.openviduService.storeChoose(
+      this.meetingService.storeChoose(
         sessionName,
         payload.sender,
         payload.receiver,
       )
-      const chooseData = this.openviduService.getChooseData(sessionName)
+
+      const chooseData = this.meetingService.getChooseData(sessionName)
       if (chooseData.length === 6) {
-        const participants = this.openviduService.getParticipants(sessionName)
-        const matches = this.openviduService.findMatchingPairs(sessionName)
+        const participants = this.meetingService.getParticipants(sessionName)
+        const matches = this.meetingService.findMatchingPairs(sessionName)
 
         const matchedPairs = matches.map(match => ({
           pair: match.pair,
@@ -195,34 +205,46 @@ export class MeetingGateway
           })
           this.cupidFlag.set(sessionName, true)
         }
+        this.meetingService.removeChooseData(sessionName)
       }
     } else {
       console.error('세션에러입니다')
     }
   }
 
+  @SubscribeMessage('startTimer')
+  handleStartTimer(client: Socket, payload: { sessionName: string }) {
+    const { sessionName } = payload
+    if (this.timerFlag.get(sessionName) == undefined) {
+      this.openviduService.startSessionTimer(sessionName, this.server)
+      this.timerFlag.set(sessionName, true)
+    }
+  }
+
   @SubscribeMessage('forwardDrawing')
-  handleFowardDrawing(
+  handleForwardDrawing(
     client: Socket,
-    payload: { userName: string; drawing: any },
+    payload: { userName: string; drawing: string; photo: string },
   ) {
-    const { drawing, userName } = payload
+    const { drawing, userName, photo } = payload
     const sessionName = this.roomid.get(userName)
+
     if (!sessionName) {
       console.error(`세션에 없는 유저이름임: ${userName}`)
       return
     }
 
-    this.openviduService.saveDrawing(sessionName, userName, drawing)
+    this.meetingService.savePhoto(sessionName, userName, photo)
+    this.meetingService.saveDrawing(sessionName, userName, drawing)
 
-    const drawings = this.openviduService.getDrawings(sessionName)
+    const drawings = this.meetingService.getDrawings(sessionName)
 
     if (Object.keys(drawings).length === 6) {
-      const participants = this.openviduService.getParticipants(sessionName)
+      const participants = this.meetingService.getParticipants(sessionName)
       participants.forEach(({ socket }) => {
         this.server.to(socket.id).emit('drawingSubmit', drawings)
       })
-      this.openviduService.resetDrawings(sessionName)
+      this.meetingService.resetDrawings(sessionName)
     }
   }
 
@@ -233,15 +255,22 @@ export class MeetingGateway
   ) {
     const { userName, votedUser } = payload
     const sessionName = this.roomid.get(userName)
-    this.openviduService.saveVote(sessionName, userName, votedUser)
 
-    const votes = this.openviduService.getVotes(sessionName)
+    this.meetingService.saveVote(sessionName, userName, votedUser)
+
+    const votes = this.meetingService.getVotes(sessionName)
 
     if (Object.keys(votes).length === 6) {
-      const winner = this.openviduService.calculateWinner(sessionName)
-      const participants = this.openviduService.getParticipants(sessionName)
+      const { winner, losers } =
+        this.meetingService.calculateWinner(sessionName)
+
+      const participants = this.meetingService.getParticipants(sessionName)
       participants.forEach(({ socket }) => {
-        this.server.to(socket.id).emit('voteResults', { winner })
+        this.server.to(socket.id).emit('voteResults', {
+          winner,
+          losers,
+          photos: this.meetingService.getPhotos(sessionName, userName),
+        })
       })
     }
   }
@@ -249,23 +278,132 @@ export class MeetingGateway
   @SubscribeMessage('winnerPrize')
   handleWinnerPrize(
     client: Socket,
-    payload: { winners: string[]; losers: string[] },
+    payload: { userName: string; winners: string[]; losers: string[] },
   ) {
-    const { winners, losers } = payload
-    const sessionName = this.roomid.get(winners[0])
-    const participants = this.openviduService.getParticipants(sessionName)
-    participants.forEach(({ socket }) => {
-      this.server
-        .to(socket.id)
-        .emit('finalResults', { winners: winners, losers: losers })
+    const { userName, winners, losers } = payload
+    const sessionName = this.roomid.get(userName)
+    const participants = this.meetingService.getParticipants(sessionName)
+
+    if (userName === winners[0])
+      participants.forEach(({ socket }) => {
+        this.server.to(socket.id).emit('finalResults', { winners, losers })
+      })
+  }
+
+  @SubscribeMessage('drawingOneToOne')
+  handleDrawingOneToOne(
+    client: Socket,
+    payload: { userName: string; winners: string[]; losers: string[] },
+  ) {
+    const { userName, winners, losers } = payload
+    let partner: string
+
+    winners.includes(userName)
+      ? (partner = winners.filter(u => u !== userName)[0])
+      : (partner = '0')
+
+    client.emit('cupidResult', {
+      lover: partner,
+      loser: losers,
     })
+  }
+
+  // 1. 10초 이내에 '1대1화상채팅하기' 버튼을 누르지 않으면 비활성
+  // 2. 성공적으로 '1대1화상채팅하기' 버튼을 눌렀을 경우 클라이언트 -> 서버(Event : chooseCam)
+  @SubscribeMessage('lastChoose')
+  handleChooseCam(
+    client: Socket,
+    payload: { sender: string; receiver: string },
+  ) {
+    // 서버 입장에서 소켓이 존재하는 방을 찾기 위함
+    const { sender, receiver } = payload
+    const sessionName = this.roomid.get(sender)
+    // 기존 정보가 있다면 새롭게 변형해서 저장할 수 있음
+    if (sessionName) {
+      this.meetingService.storeChoose(sessionName, sender, receiver)
+      // 방과 일치하는 매칭결과 정보 가져오기
+      const chooseData = this.meetingService.getChooseData(sessionName)
+      if (chooseData.length === 6) {
+        // 방과 일치하는 참여자 정보 가져오기
+        const participant = this.meetingService.getParticipants(sessionName)
+        // 매칭된 쌍의 정보를 가지고 있음
+        // [
+        // { pair: [ 'Alice', 'Bob' ] },
+        // { pair: [ 'Charlie', 'David' ] },
+        // { pair: [ 'Eve', 'Frank' ] }
+        // ]
+        const matches = this.meetingService.findMatchingPairs(sessionName)
+
+        if (this.lastCupidFlag.get(sessionName) == undefined) {
+          participant.forEach(({ socket, name }) => {
+            const matchedPair = matches.find(elem => elem.pair.includes(name))
+            if (matchedPair) {
+              const partner = matchedPair.pair.find(elem => elem !== name)
+              this.server.to(socket.id).emit('matching', { lover: partner })
+            } else {
+              this.server.to(socket.id).emit('matching', { lover: '0' })
+            }
+
+            this.server.to(socket.id).emit('lastChooseResult', chooseData)
+          })
+          this.lastCupidFlag.set(sessionName, true)
+        }
+      }
+    } else {
+      console.error('세션이 존재하지 않습니다.')
+    }
+  }
+
+  @SubscribeMessage('moveToPrivateRoom')
+  async handleMoveToPrivateRoom(
+    client: Socket,
+    payload: { sessionName: string; myName: string; partnerName: string },
+  ) {
+    const { sessionName, myName, partnerName } = payload
+    const participant = this.meetingService.getParticipants(sessionName)
+    if (this.acceptanceStatus[partnerName] === true) {
+      const newSessionName = `${myName}-${partnerName}`
+
+      const newSession = await this.meetingService.createSession(newSessionName)
+
+      const partner = await participant.find(
+        participant => participant.name === partnerName,
+      )
+      this.meetingService.addParticipant(newSessionName, myName, client)
+      this.meetingService.addParticipant(
+        newSessionName,
+        partnerName,
+        partner.socket,
+      )
+
+      const enterToken =
+        await this.meetingService.generateTokens(newSessionName)
+
+      const myToken = enterToken.find(elem => elem.participant === myName).token
+      const partnerToken = enterToken.find(
+        elem => elem.participant === partnerName,
+      ).token
+
+      if (myToken && partnerToken) {
+        this.server
+          .to(client.id)
+          .emit('choice', { sessionId: newSessionName, token: myToken })
+        this.server
+          .to(partner.socket.id)
+          .emit('choice', { sessionId: newSessionName, token: partnerToken })
+      } else {
+        console.error('방 생성 실패!')
+      }
+    } else {
+      this.acceptanceStatus[myName] = true
+    }
   }
 
   @SubscribeMessage('leave')
   handleLeave(client: Socket, payload: { participantName }) {
     const sessionName = this.roomid.get(payload.participantName)
     if (sessionName) {
-      this.openviduService.removeParticipant(
+      this.meetingService.removeParticipant(
         sessionName,
         client,
         payload.participantName,
@@ -275,5 +413,19 @@ export class MeetingGateway
     this.cupidFlag.delete(sessionName)
     delete this.connectedUsers[payload.participantName]
     delete this.connectedSockets[client.id]
+    this.timerFlag.delete(sessionName)
+  }
+
+  @SubscribeMessage('emoji')
+  handleEmoji(
+    client: Socket,
+    payload: { nickname: string; emojiIndex: string },
+  ) {
+    const { nickname, emojiIndex } = payload
+    const sessionName = this.roomid.get(nickname)
+    const participants = this.meetingService.getParticipants(sessionName)
+    participants.forEach(({ socket }) => {
+      this.server.to(socket.id).emit('emojiBroadcast', { nickname, emojiIndex })
+    })
   }
 }
